improved typing
