+++ src/lib/index.ts
@@ -65,7 +65,7 @@ export { P5Asciifier } from './Asciifier';
 export { P5AsciifyColorPalette } from './ColorPalette';
 export { P5AsciifyGrid } from './Grid';
 export { P5AsciifyFontManager } from './FontManager';
-export { P5AsciifyHookManager, type HookType } from './HookManager';
+export { P5AsciifyHookManager } from './HookManager';
 export * from './types';
 
 if (typeof window !== 'undefined') {
